pyproject.toml
[project]
name = "sushi-ordering-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.8",
    "django-phonenumber-field>=8.3.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-xml>=2.0.0",
    "djoser>=2.3.3",
    "phonenumbers>=9.0.18",
]
